+++ js/dterm.js
@@ -10,7 +10,7 @@
  * Released under the MIT license
  *
  */
-/* global jQuery */
+/* global jQuery setTimeout IntersectionObserver */
 (function($) {
     $.extend_if_has = function(desc, source, array) {
         for (var i = array.length; i--;) {
@@ -39,14 +39,29 @@
             };
         }
         var self = this;
+        if (window.IntersectionObserver) {
+            var visibility_observer = new IntersectionObserver(function() {
+                if (self.is(':visible')) {
+                    terminal.enable().resize();
+                } else {
+                    self.disable();
+                }
+            }, {
+                root: document.body
+            });
+            visibility_observer.observe(terminal[0]);
+        }
         this.dialog($.extend(options, {
             resizeStop: function() {
                 var content = self.find('.ui-dialog-content');
                 terminal.resize(content.width(), content.height());
             },
             open: function() {
-                terminal.focus();
-                terminal.resize();
+                if (!window.IntersectionObserver) {
+                    setTimeout(function() {
+                        terminal.enable().resize();
+                    }, 100);
+                }
             },
             show: 'fade',
             closeOnEscape: false